custom-elements/connected-callbacks-html-fragment-parsing.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: connectedCallback not called during HTML fragment parsing</title>
<meta name="author" title="Rob Buis" href="mailto:rbuis@igalia.com">
<meta name="assert" content="connectedCallback must not be called during HTML fragment parsing">
<link rel="help" href="https://html.spec.whatwg.org/multipage/parsing.html#create-an-element-for-the-token">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/custom-elements-helpers.js"></script>
</head>
<body>
<div id="log"></div>
<script>

var iteration = 0;

document_types().forEach(function (entry) {
    let documentName = entry.name;
    let getDocument = entry.create;
    let calls = 0;

    promise_test(function () {
        class Parenter extends HTMLElement {
            connectedCallback() {
                const child = this.firstChild;
                this.removeChild(child);
                this.appendChild(child);
            }
        }
        class Child extends HTMLElement {
            connectedCallback() { calls++; }
        }
        let parenter = 'x-parenter' + iteration;
        let child = 'x-child' + iteration++;
        customElements.define(parenter, Parenter);
        customElements.define(child, Child);
        return getDocument().then(function (doc) {
            document.documentElement.innerHTML = '<' + parenter + '><' + child + '></' + child + '></' + parenter + '>';
            doc.documentElement.appendChild(document.documentElement.firstChild);
            assert_equals(calls, 1);
        });
    }, 'Inserting a custom element into ' + documentName + ' using HTML fragment parsing must not call connectedCallback.');
});

</script>
</body>
</html>
